Fix integral division widening to float warnings
In this commit, the new lint warning introduced with Scala 2.12.18, which pertains to integer divisions that are implicitly converted to floats, is addressed.

+++ netlogo-gui/src/main/app/tools/AgentMonitorViewPanel.scala
@@ -18,7 +18,7 @@ class AgentMonitorViewPanel(workspace: GUIWorkspace) extends JPanel(new BorderLa
 
   locally {
     add(view, BorderLayout.CENTER)
-    view.setSize(workspace.world.worldWidth, workspace.world.worldHeight, 255 / workspace.world.worldWidth)
+    view.setSize(workspace.world.worldWidth, workspace.world.worldHeight, 255.toDouble / workspace.world.worldWidth)
     view.applyNewFontSize(workspace.view.fontSize, 0)
     watchButton.setFocusable(false)
     val controls = new JPanel
